+++ app/admin/adminCtrl.js
@@ -46,9 +46,11 @@
                 if ($routeParams.section) {
                     vm.navState.screen = $routeParams.section;
                 }
+                if ($routeParams.subSection) {
+                    vm.navState[vm.navState.screen] = $routeParams.subSection;
+                }
                 if ($routeParams.productId)  {
                     vm.navState.reports = '';
-                    vm.navState.dpManagement = 'manage';
                     vm.productId = $routeParams.productId;
                 }
 
@@ -111,6 +113,15 @@
                     path = path.substring(0,path.lastIndexOf('/'));
                     $location.path(path);
                 }
+                clearSubsection();
+            }
+
+            function clearSubsection () {
+                var path = $location.path();
+                if ($routeParams.productId)  {
+                    path = path.substring(0,path.lastIndexOf('/'));
+                    $location.path(path);
+                }
             }
 
             function getUsername () {
@@ -136,7 +147,7 @@
             function isCmsStaff () {
                 return authService.isCmsStaff();
             }
-            
+
             function isOncStaff() {
             	return authService.isOncStaff();
             }